+++ batch_priority/urls.py
@@ -14,6 +14,7 @@ urlpatterns = [
     path("locations/", views.location, name="locations"),
     path("locations/<int:batch_id>/", views.locations, name="location"),
     path("products/", views.product_list, name="product_list"),
+    path('wip-queue/', views.WIPQueueView.as_view(), name='wip_queue'),
     path("batch_history/<int:batch_id>/", views.batch_history, name="batch_history"),
     path("batch/<int:batch_id>/", views.batch_detail, name="batch_detail"),
     path("reports/completed_on/", views.CompletedOnView.as_view(), name="completed_on"),